Text2sql/backend/requirements.txt
fastapi>=0.109.2
uvicorn[standard]>=0.27.1
python-jose>=3.3.0
python-multipart>=0.0.9
sqlalchemy>=2.0.27
pydantic>=2.6.1
pydantic-settings>=2.1.0
pandas>=2.2.0
openpyxl>=3.1.2
google-genai>=0.3.0